wisefin/requirements.txt
APScheduler==3.6.3
asgiref==3.2.7
atomicwrites==1.4.0
attrs==19.3.0
awscli==1.18.69
boto3==1.13.15
boto3-stubs==1.13.18.0
botocore==1.16.19
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
Django==3.0.4
django-apscheduler==0.4.0
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.2.1
django-memcached==0.1.2
djangorestframework==3.11.0
djangorestframework-stubs==1.1.0
docutils==0.15.2
env.py==0.2.2
environs==8.0.0
html5lib==1.0.1
idna==2.9
importlib-metadata==1.7.0
jmespath==0.10.0
marshmallow==3.7.1
more-itertools==8.4.0
mysql-connector==2.2.9
mysqlclient==1.4.6
numpy==1.19.1
packaging==20.4
pandas==0.25.3
pdfkit==0.6.1
Pillow==7.1.2
pluggy==0.13.1
py==1.9.0
py-sql==1.0.0
pyasn1==0.4.8
pyparsing==2.4.7
PyPDF2==1.26.0
pytest==5.4.3
python-dateutil==2.8.1
python-dotenv==0.14.0
python-memcached==1.59
pytz==2019.3
PyYAML==5.3.1
reportlab==3.5.42
requests==2.23.0
rsa==3.4.2
s3transfer==0.3.3
schedule==0.6.0
six==1.15.0
sqlparse==0.3.1
tzlocal==2.1
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
xhtml2pdf==0.2.4
zipp==3.1.0
django-rest-knox==4.1.0
num2words==0.5.10
camelcase==0.2
python-barcode
XlsxWriter
httpagentparser==1.9.1
xlrd==1.2.0
openpyxl==3.0.7
django-csp==3.7
django-ipware==4.0.2